requirements.txt
streamlit
pandas
torch
transformers
supabase
 dev
fpdf
streamlit-cropper
numpy
requests
Pillow

storage3
python-dateutil
fpdf
requests
openai-clip


 main

